Add an experimental Makefile
Add an experimental Makefile to the project. Since this project is
based on the IDE Qt-Creator (which uses qmake for building) it is
difficult to compile it without qmake. This makefile is intended to fix
this by offering the option to compile the project with the GNU
Toolchain using make and g++.